requirements.txt
psutil
yara-python
flask

